fine_tuning/requirements.txt
# Phi-4 Fine-Tuning Requirements
# Install: pip install -r requirements.txt
# Note: Install torch with CUDA first:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121

# Core training
unsloth[colab-new]>=2024.12.0
transformers>=4.47.0
datasets>=3.1.0
peft>=0.14.0
accelerate>=1.2.0
bitsandbytes>=0.45.0
trl>=0.15.0
vllm>=0.6.0
# install separately: pip install flash-attn --no-build-isolation

# Model utilities
sentencepiece>=0.2.0
protobuf>=4.25.0
tokenizers>=0.21.0

# Training monitoring
tensorboard>=2.18.0
wandb>=0.19.0

# Data processing
pyyaml>=6.0.2
tqdm>=4.67.0
numpy>=1.26.0
pandas>=2.2.0
rich>=13.9.0

# Quantization (deployment)
llama-cpp-python>=0.3.0
auto-gptq>=0.7.1
optimum>=1.24.0

# Evaluation
openai>=1.58.0
anthropic>=0.40.0
rouge-score>=0.1.2
nltk>=3.9.0
jsonlines>=4.0.0
